backend/pyproject.toml
[tool.poetry]
name = "poke-collect"
version = "0.1.0"
description = "Organise your Pokemon cards and sealed item collection"
authors = ["Thomas PERRAIS"]
readme = "README.md"
packages = [{include = "app", from="."}]

[tool.poetry.dependencies]
python = "3.12.*"
fastapi = "^0.118.3"
uvicorn = "^0.37.0"
sqlalchemy = "^2.0.44"
strawberry-graphql = {extras = ["fastapi"], version = "^0.283.2"}
mypy = "^1.18.2"
sqlalchemy-enum34 = "^2.0.0"
pymysql = "^1.1.2"
alembic = "^1.17.1"
cryptography = "^46.0.3"
pandas = "^2.3.3"
requests = "^2.32.5"
unidecode = "^1.4.0"

[[tool.poetry.source]]
name = "pypi"
priority = "primary"

[tool.poetry.group.dev.dependencies]
mypy-boto3-s3 = "^1.38.0"
black = "^23.11.0"
flake8 = "^6.1.0"
pytype = "^2024.1.24"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
